Resolve documentation warnings
