+++ Softeq.XToolkit.Bindings.iOS/ObservableGroupTableViewSource.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Collections.Specialized;
 using System.Linq;
 using Foundation;
 using Softeq.XToolkit.Bindings.iOS.Extensions;
