+++ Softeq.XToolkit.Bindings/Binding.cs
@@ -21,6 +21,11 @@ namespace Softeq.XToolkit.Bindings
         protected WeakReference? TopTarget;
 
         /// <summary>
+        ///     Occurs when the value of the databound property changes.
+        /// </summary>
+        public abstract event EventHandler ValueChanged;
+
+        /// <summary>
         ///     Gets or sets the mode of the binding. OneTime means that the target property will be set once (when the binding is
         ///     created) but that subsequent changes will be ignored. OneWay means that the target property will be set, and
         ///     if the PropertyChanged event is raised by the source, the target property will be updated. TwoWay means that the
@@ -61,10 +66,5 @@ namespace Softeq.XToolkit.Bindings
         ///     be set to the target value.
         /// </summary>
         public abstract void ForceUpdateValueFromTargetToSource();
-
-        /// <summary>
-        ///     Occurs when the value of the databound property changes.
-        /// </summary>
-        public abstract event EventHandler ValueChanged;
     }
 }
