+++ Softeq.XToolkit.Common/Commands/ICommand.cs
@@ -12,7 +12,7 @@ namespace Softeq.XToolkit.Common.Commands
     public interface ICommand<in T> : ICommand
     {
         /// <summary>
-        ///     Provides a strongly-typed variant of <see cref="ICommand.CanExecute(object)"/>.
+        ///     Provides a strongly-typed variant of <see cref="T:System.Windows.Input.ICommand.CanExecute(object)"/>.
         /// </summary>
         /// <param name="parameter">The input parameter.</param>
         /// <returns>Whether or not the current command can be executed.</returns>
@@ -20,7 +20,7 @@ namespace Softeq.XToolkit.Common.Commands
         bool CanExecute(T parameter);
 
         /// <summary>
-        ///     Provides a strongly-typed variant of <see cref="ICommand.Execute(object)"/>.
+        ///     Provides a strongly-typed variant of <see cref="T:System.Windows.Input.ICommand.Execute(object)"/>.
         /// </summary>
         /// <param name="parameter">The input parameter.</param>
         /// <remarks>Use this overload to avoid boxing, if <typeparamref name="T"/> is a value type.</remarks>
