+++ Softeq.XToolkit.Common/Threading/IMainThreadExecutor.cs
@@ -26,6 +26,7 @@ namespace Softeq.XToolkit.Common.Threading
         ///     Invokes asynchronously the specified code on the main UI thread.
         /// </summary>
         /// <param name="action">The action to execute.</param>
+        /// <returns>A <see cref="T:System.Threading.Tasks.Task"/> representing the asynchronous operation.</returns>
         Task OnUIThreadAsync(Action action);
 
         /// <summary>
