+++ Softeq.XToolkit.PushNotifications.iOS/INotificationsPermissionsService.cs
@@ -8,9 +8,9 @@ namespace Softeq.XToolkit.PushNotifications.iOS
     public interface INotificationsPermissionsService
     {
         /// <summary>
-        ///     Requests push notifications permissions
+        ///     Requests push notifications permissions.
         /// </summary>
-        /// <returns>Task with true value if permissions were granted, and false value otherwise</returns>
+        /// <returns>Task with true value if permissions were granted, and false value otherwise.</returns>
         Task<bool> RequestNotificationsPermissions();
     }
 }
